dosyalar debug edildi ve toplandı

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: kali <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/01 15:49:52 by kali              #+#    #+#             */
-/*   Updated: 2023/03/01 15:55:52 by kali             ###   ########.fr       */
+/*   Updated: 2023/03/02 17:55:35 by kali             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	long unsigned int	i;
 	int	res1;
 	int	res2;
